Added: EVMProvider reusable class
